+++ integration-tests/fixtures/infinite-loop-mod/control.ts
@@ -1,0 +1,3 @@
+if ("factorio-test" in script.active_mods) {
+  require("__factorio-test__/init")(["tests"])
+}